refactor(procfs): split parsers module
